Fixes #11160 : Increase the number of times the scripts attempts to call cpinit before failing.
The prevailing thought is that this is due to lack of entropy on vms. If that is trye, all that can be done
is wait longer. This bumps up the script to attempt to connect 40 times, waiting 20 seconds between attempts.
This should give 800 seconds for the machine to get entropy