Pre-create some tomcat config files in selinux enabled acceptance tests
Works around https://github.com/theforeman/puppet-candlepin/issues/185
When the tomcat RPM is installed, matchpathcon returns
different file context before vs. after touching the files
/usr/share/tomcat/conf/{login.config,conf.d/jaas.conf} and
/usr/share/tomcat/conf/cert-{users,roles}.properties
which breaks idempotence in selinux enabled environments.
This bug should likely be fixed in the RPM packaging of tomcat, but
in the meantime we can work around it by pre-creating the files to
ensure correct file context is returned when puppet is first run.