Disable FIPS support within Tomcat Java environment
On EL8 FIPS enabled systems, Candlepin cannot run with FIPS enabled
within the Java stack. The solution is to turn off FIPS within the Tomcat
environment but leave it enabled for the system itself. This mirrors
the behavior on EL7 FIPS mode and is required until Candlepin can
adopt to working natively in a FIPS enabled Java environment.