Fixes #35638 - Update ciphers to be SHA256 or SHA384
In 832bafa66c9fdc8d632908613695691e90f78583 the list was created in the
initial commit. Then c5a36f728cc12443709d0437b205c4a9e32c0fbe made it a
parameter, but didn't change the values. Since 2013 there are stronger
(non-SHA1) ciphers. This is important since the FUTURE crypto policy
disallows SHA1 ciphers.

The old ciphers are less secure. In our setup clients talk to Foreman
and Foreman talks to Candlepin so this should be safe in terms of
compatibility with older clients.