Rewrite _candlepin_database.conf from ERB to EPP
Like the tomcat changes, this also implements better secret handling.
This bumps the required PostgreSQL module version to one that deals with
Sensitive. puppetlabs-mysql is a soft dependency, but now 11.1.0 is the
oldest compatible version.

The specs are modified because .without_content(senstive()) doesn't
work. Instead it now checks for the line ending to ensure it's rendered
correctly.