Avoid dependency cycles with glibc-langpack-en
By calling stdlib::ensure_packages() in the class itself the resource
gets tied to the class. If other packages also depend on the resource,
you can get dependency cycles.

In particular, foreman also does this and that causes a dependency
cycle.

By using a class that is included the resource "floats" in the catalog.
We then only ensure the package is installed somewhere before we create
the database.

Fixes: ca7714557135 ("Ensure glibc-langpack-en is always installed")